Add authentitaction, authorization, change views
- authorization for add, move, moveup, edit and delete actions
- change views
- refactor controller

+++ app/Http/Controllers/CategoryController.php
@@ -5,12 +5,15 @@ declare(strict_types=1);
 namespace App\Http\Controllers;
 
 use App\Models\Category;
+use App\Models\User;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
-use Symfony\Component\Console\Input\Input;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\View\View;
 
 class CategoryController extends Controller
 {
-    public function index(Request $request)
+    public function index(Request $request): View
     {
         $this->validate($request, [
             'show' => 'nullable|integer'
@@ -23,31 +26,39 @@ class CategoryController extends Controller
 
         return view('main', [
             'categories' => $categories,
-            'allCategories' => $allCategories,
-            'showCategory' => $branchId === 0
-                ? 'Root'
-                : Category::where('id', '=', $branchId)
-                    ->firstOrFail()
-                    ->title
+            'allCategories' => $allCategories
         ]);
     }
 
-    public function store(Request $request)
+    public function store(Request $request, User $user): RedirectResponse
     {
+        if (! Gate::allows('admin', $user)) {
+            abort(403);
+        }
+
         $this->validate($request, [
-            'title' => 'required|alpha_num|max:50'
+            'title' => 'required|alpha_num|max:50',
+            'parent_id' => 'required|integer'
         ]);
 
-        $input = $request->all();
-        $input['parent_id'] = empty($input['parent_id']) ? 0 : $input['parent_id'];
+        if(Category::where('id', '=', $request->parent_id)->doesntExist() && $request->parent_id != 0) {
+            return back()->with('error', 'Nie ma takiego rodzica');
+        }
 
-        Category::create($input);
+        Category::create([
+            'title' => $request->title,
+            'parent_id' => $request->parent_id
+        ]);
 
-        return back()->with('success', 'New Category added successfully');
+        return back()->with('success', 'Nowa kategoria została dodana');
     }
 
-    public function moveUp(Request $request)
+    public function moveUp(Request $request, User $user): RedirectResponse
     {
+        if (! Gate::allows('admin', $user)) {
+            abort(403);
+        }
+
         $this->validate($request, [
             'id' => 'required|integer',
             'parent_id' => 'required|integer'
@@ -57,50 +68,78 @@ class CategoryController extends Controller
 
         Category::where('id', '=', $request->id)->update(['parent_id' => $parent->parent_id]);
 
-        return back()->with('success', 'Moved to upper category');
+        return back()->with('success', 'Przeniesiono poziom wyżej');
     }
 
-    public function move(Request $request)
+    public function move(Request $request, User $user): RedirectResponse
     {
+        if (! Gate::allows('admin', $user)) {
+            abort(403);
+        }
+
         $this->validate($request, [
-            'category_id' => 'required|integer',
+            'id' => 'required|integer',
             'parent_id' => 'required|integer'
         ]);
 
-        if ($request->category_id !== $request->parent_id) {
-            Category::where('id', '=', $request->category_id)
-                ->update(['parent_id' => $request->parent_id]);
-
-            return back()->with('success', 'Moved to other category');
+        if(Category::where('id', '=', $request->id)->doesntExist() && $request->id != 0) {
+            return back()->with('error', 'Nie ma takiej kategorii');
         }
 
-        return back()->with('error', 'Can\'t move to the same category and parent');
+        if(Category::where('id', '=', $request->parent_id)->doesntExist() && $request->parent_id != 0) {
+            return back()->with('error', 'Nie ma takiego rodzica');
+        }
+
+        if ($request->id === $request->parent_id) {
+            return back()->with('error', 'Nie można przenieść kategorii do niej samej');
+        }
+
+        Category::where('id', '=', $request->id)
+            ->update(['parent_id' => $request->parent_id]);
+
+        return back()->with('success', 'Przeniesiono do innej gałęzi');
     }
 
-    public function destroy(Request $request)
+    public function destroy(Request $request, User $user): RedirectResponse
     {
+        if (! Gate::allows('admin', $user)) {
+            abort(403);
+        }
+
         $this->validate($request, [
             'id' => 'required|integer'
         ]);
 
+        if(Category::where('id', '=', $request->id)->doesntExist() && $request->id != 0) {
+            return back()->with('error', 'Nie ma takiej kategorii');
+        }
+
         $this->delete((int) $request->id);
 
-        return back()->with('success', 'Deleted selected category');
+        return back()->with('success', 'Usunięto kategorię');
     }
 
-    public function update(Request $request)
+    public function update(Request $request, User $user): RedirectResponse
     {
+        if (! Gate::allows('admin', $user)) {
+            abort(403);
+        }
+
         $this->validate($request, [
             'id' => 'required|integer',
             'newTitle' => 'required|alpha_num|max:50'
         ]);
 
+        if(Category::where('id', '=', $request->id)->doesntExist() && $request->id != 0) {
+            return back()->with('error', 'Nie ma takiej kategorii');
+        }
+
         Category::find($request->id)->update(['title' => $request->newTitle]);
 
-        return back()->with('success', 'Edited selected category');
+        return back()->with('success', 'Zmieniono nazwę kategorii');
     }
 
-    private function delete(int $id)
+    private function delete(int $id): void
     {
         Category::where('id', '=', $id)->delete();
         $ids = Category::where('parent_id', '=', $id)->get('id')->toArray();
